Added index.html redirect to P1/index.html
Currently our homepage is in P1/index.html and it will not load if the website is accessed with a URL without a path to P1/index.html (i.e. https://green-market.ml would 404). The index.html file in this commit would just serves as a redirect to our real homepage which would allow us to keep the same project structure.